+++ CHANGELOG.md
@@ -7,6 +7,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [2.1.0] - 2024-08-09
+### Added
+- Added User32 features needed to map virtual keys to char.
 ### Fixed
 - Fixed namespace of some structures.